test: 그외 테스트 레거시 코드
