+++ backend/src/test/java/com/pickeat/backend/acceptance_test/piece/wish/WishPicturePieceTest.java
