+++ backend/src/test/java/com/pickeat/backend/acceptance_test/scenario/PickeatByTemplateScenarioTest.java
@@ -1,4 +1,4 @@
-package com.pickeat.backend.acceptance_test.scenario;
+package com.pickeat.backend.tobe.acceptance_test.scenario;
 
 import static com.pickeat.backend.acceptance_test.piece.participant.ParticipantPieceTest.참가자_생성;
 import static com.pickeat.backend.acceptance_test.piece.participant.ParticipantPieceTest.참가자_선택_완료_표시;
@@ -9,12 +9,12 @@ import static com.pickeat.backend.acceptance_test.piece.pickeat.PickeatPieceTest
 import static com.pickeat.backend.acceptance_test.piece.pickeat.PickeatPieceTest.픽잇_정보_조회;
 import static com.pickeat.backend.acceptance_test.piece.pickeat.PickeatPieceTest.픽잇_활성화_상태_조회;
 import static com.pickeat.backend.acceptance_test.piece.pickeat.PickeatPieceTest.픽잇의_참가자_요약_정보_조회;
-import static com.pickeat.backend.acceptance_test.piece.restaurant.RestaurantPieceTest.식당_소거;
-import static com.pickeat.backend.acceptance_test.piece.restaurant.RestaurantPieceTest.식당_좋아요;
-import static com.pickeat.backend.acceptance_test.piece.restaurant.RestaurantPieceTest.식당_좋아요_취소;
-import static com.pickeat.backend.acceptance_test.piece.restaurant.RestaurantPieceTest.위시리스트_기반으로_식당_생성;
-import static com.pickeat.backend.acceptance_test.piece.restaurant.RestaurantPieceTest.픽잇의_식당_조회;
-import static com.pickeat.backend.acceptance_test.piece.wish.WishListPieceTest.템플릿_위시리스트_조회;
+import static com.pickeat.backend.tobe.acceptance_test.piece.restaurant.RestaurantPieceTest.식당_제외;
+import static com.pickeat.backend.tobe.acceptance_test.piece.restaurant.RestaurantPieceTest.식당_좋아요;
+import static com.pickeat.backend.tobe.acceptance_test.piece.restaurant.RestaurantPieceTest.식당_좋아요_취소;
+import static com.pickeat.backend.tobe.acceptance_test.piece.restaurant.RestaurantPieceTest.템플릿_기반_식당_목록_생성;
+import static com.pickeat.backend.tobe.acceptance_test.piece.restaurant.RestaurantPieceTest.픽잇의_식당_조회;
+import static com.pickeat.backend.tobe.acceptance_test.piece.template.TemplatePieceTest.템플릿_목록_조회;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
@@ -26,10 +26,10 @@ import com.pickeat.backend.pickeat.application.dto.response.ParticipantStateResp
 import com.pickeat.backend.pickeat.application.dto.response.PickeatResponse;
 import com.pickeat.backend.pickeat.application.dto.response.PickeatStateResponse;
 import com.pickeat.backend.restaurant.application.dto.request.RestaurantExcludeRequest;
-import com.pickeat.backend.restaurant.application.dto.request.WishRestaurantRequest;
+import com.pickeat.backend.restaurant.application.dto.request.TemplateRestaurantRequest;
 import com.pickeat.backend.restaurant.application.dto.response.RestaurantResponse;
 import com.pickeat.backend.restaurant.application.dto.response.RestaurantResultResponse;
-import com.pickeat.backend.wish.application.dto.response.WishListResponse;
+import com.pickeat.backend.template.application.dto.response.TemplateResponse;
 import io.restassured.RestAssured;
 import java.util.List;
 import org.junit.jupiter.api.AfterEach;
@@ -40,7 +40,7 @@ import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.jdbc.Sql;
 
-@Sql("/init/template_data.sql")
+@Sql("/init/template_data_v2.sql")
 @DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_CLASS)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class PickeatByTemplateScenarioTest {
@@ -59,13 +59,13 @@ public class PickeatByTemplateScenarioTest {
     }
 
     @Test
-    void 위치_기반_픽잇_플로우() {
+    void 텤플릿_기반_픽잇_플로우() {
         // 템플릿 조회
-        List<WishListResponse> templates = 템플릿_위시리스트_조회();
+        List<TemplateResponse> templates = 템플릿_목록_조회();
 
         // 픽잇 생성
         PickeatResponse createdPickeat = 외부용_픽잇_생성(new PickeatRequest("우테코 점심 픽잇"));
-        위시리스트_기반으로_식당_생성(createdPickeat.code(), new WishRestaurantRequest(templates.getFirst().id()));
+        템플릿_기반_식당_목록_생성(createdPickeat.code(), new TemplateRestaurantRequest(templates.getFirst().id()));
 
         // 참여자 생성
         PickeatResponse pickeat = 픽잇_정보_조회(createdPickeat.code());
@@ -74,13 +74,13 @@ public class PickeatByTemplateScenarioTest {
         TokenResponse participant3Token = 참가자_생성(new ParticipantRequest("참여자3", pickeat.id()));
 
         // 참여자들의 식당 소거
-        List<RestaurantResponse> restaurants = 픽잇의_식당_조회(pickeat.code(), participant1Token.token(), null);
+        List<RestaurantResponse> restaurants = 픽잇의_식당_조회(pickeat.code(), null, participant1Token.token());
         List<Long> restaurantIds = restaurants.stream().map(RestaurantResponse::id).toList();
-        식당_소거(new RestaurantExcludeRequest(restaurantIds.subList(0, 1)), participant1Token.token());
-        식당_소거(new RestaurantExcludeRequest(restaurantIds.subList(1, 2)), participant2Token.token());
-        식당_소거(new RestaurantExcludeRequest(restaurantIds.subList(2, 3)), participant3Token.token());
+        식당_제외(new RestaurantExcludeRequest(restaurantIds.subList(0, 1)), participant1Token.token());
+        식당_제외(new RestaurantExcludeRequest(restaurantIds.subList(1, 2)), participant2Token.token());
+        식당_제외(new RestaurantExcludeRequest(restaurantIds.subList(2, 3)), participant3Token.token());
 
-        List<RestaurantResponse> restaurantAfterExcluded = 픽잇의_식당_조회(pickeat.code(), participant1Token.token(), null);
+        List<RestaurantResponse> restaurantAfterExcluded = 픽잇의_식당_조회(pickeat.code(), null, participant1Token.token());
         checkExcludedRestaurants(restaurantIds.subList(0, 3), restaurantAfterExcluded);
 
         참가자_선택_완료_표시(participant1Token.token());
@@ -91,7 +91,7 @@ public class PickeatByTemplateScenarioTest {
         ParticipantStateResponse participantStateSummary = 픽잇의_참가자_요약_정보_조회(pickeat.code());
         checkParticipantState(participantStateSummary, 3, 3);
 
-        List<RestaurantResponse> noneExcludedRestaurant = 픽잇의_식당_조회(pickeat.code(), participant1Token.token(), false);
+        List<RestaurantResponse> noneExcludedRestaurant = 픽잇의_식당_조회(pickeat.code(), false, participant1Token.token());
         식당_좋아요(noneExcludedRestaurant.get(0).id(), participant1Token.token());
         식당_좋아요(noneExcludedRestaurant.get(1).id(), participant2Token.token());
         식당_좋아요(noneExcludedRestaurant.get(1).id(), participant3Token.token());
