+++ backend/src/test/java/com/pickeat/backend/fake/restaurant/TestKakaoMapClientConfig.java
@@ -1,4 +1,4 @@
-package com.pickeat.backend.fake.restaurant;
+package com.pickeat.backend.tobe.fake.restaurant;
 
 import com.pickeat.backend.restaurant.application.RestaurantSearchClient;
 import org.springframework.context.annotation.Bean;
@@ -6,10 +6,10 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 
 @Profile("test")
-@Configuration
+@Configuration("TestKakaoMapClientConfigV2")
 public class TestKakaoMapClientConfig {
 
-    @Bean
+    @Bean("RestaurantSearchClientV2")
     public RestaurantSearchClient restaurantSearchClient() {
         return new FakeRestaurantSearchClient();
     }
