+++ backend/src/test/java/com/pickeat/backend/fake/wish/TestImageUploadClientConfiguration.java
@@ -1,4 +1,4 @@
-package com.pickeat.backend.fake.wish;
+package com.pickeat.backend.tobe.fake.wish;
 
 import com.pickeat.backend.wish.application.ImageUploadClient;
 import org.springframework.beans.factory.annotation.Value;
@@ -7,10 +7,10 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 
 @Profile("test")
-@Configuration
+@Configuration("TestImageUploadClientConfigurationV2")
 public class TestImageUploadClientConfiguration {
 
-    @Bean
+    @Bean("ImageUploadClientV2")
     public ImageUploadClient localImageUploadClient(
             @Value("${default.wish.image.url}") String defaultImageUrl,
             @Value("${external.s3.wish.image.key.prefix}") String keyPrefix
