+++ backend/src/test/java/com/pickeat/backend/fixture/WishFixture.java
@@ -1,19 +1,21 @@
-package com.pickeat.backend.fixture;
+package com.pickeat.backend.tobe.fixture;
 
-import com.pickeat.backend.restaurant.domain.FoodCategory;
+import com.pickeat.backend.restaurant.domain.Picture;
+import com.pickeat.backend.restaurant.domain.RestaurantInfo;
+import com.pickeat.backend.room.domain.Room;
 import com.pickeat.backend.wish.domain.Wish;
-import com.pickeat.backend.wish.domain.WishList;
 
 public class WishFixture {
 
-    public static Wish create(WishList wishList) {
-        return new Wish(
-                "위시",
-                FoodCategory.KOREAN,
-                "도로명주소",
-                "태그1,태그2",
-                "https://place.map.kakao.com/505348601",
-                wishList
-        );
+    public static Wish create(Room room) {
+        return new Wish(room, RestaurantInfoFixture.create("test_restaurant"));
+    }
+
+    public static Wish create(Room room, RestaurantInfo restaurantInfo) {
+        return new Wish(room, restaurantInfo);
+    }
+
+    public static Wish create(Room room, Picture picture) {
+        return new Wish(room, RestaurantInfoFixture.create("test_restaurant", picture));
     }
 }
